% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchBASSAA}
\alias{launchBASSAA}
\title{Launch BASSAA}
\usage{
launchBASSAA(...)
}
\value{
Shiny application.
}
\description{
Launch BASSAA
}
\details{
Launch BASSAA
}
\keyword{Acute}
\keyword{Ankylosing}
\keyword{Anterior}
\keyword{Atlas}
\keyword{B27}
\keyword{PBMC}
\keyword{SDA}
\keyword{Spondylitis}
\keyword{Uveitis}
\keyword{and}
\keyword{shiny}
